+++ circuits/src/instructions/base64.rs
@@ -82,6 +82,53 @@ impl<F: PrimeField, const M: usize, const A: usize> From<Base64Vec<F, M, A>>
     }
 }
 
+#[cfg(feature = "extraction")]
+pub mod extraction {
+    //! Extraction specific logic related to the base64 vector.
+
+    use extractor_support::cells::{
+        ctx::{ICtx, LayoutAdaptor, OCtx},
+        load::LoadFromCells,
+        store::StoreIntoCells,
+        CellReprSize,
+    };
+    use ff::PrimeField;
+    use midnight_proofs::{plonk::Error, ExtractionSupport};
+
+    use super::Base64Vec;
+    use crate::{types::AssignedByte, utils::extraction::IR, vec::AssignedVector};
+
+    impl<F: PrimeField, const M: usize, const A: usize> CellReprSize for Base64Vec<F, M, A> {
+        const SIZE: usize = <AssignedVector<F, AssignedByte<F>, M, A> as CellReprSize>::SIZE;
+    }
+
+    impl<F: PrimeField, C, const M: usize, const A: usize, L>
+        LoadFromCells<F, C, ExtractionSupport, L> for Base64Vec<F, M, A>
+    {
+        fn load(
+            ctx: &mut ICtx<F, ExtractionSupport>,
+            chip: &C,
+            layouter: &mut impl LayoutAdaptor<F, ExtractionSupport, Adaptee = L>,
+            injected_ir: &mut IR<F>,
+        ) -> Result<Self, Error> {
+            ctx.load(chip, layouter, injected_ir).map(Self)
+        }
+    }
+    impl<F: PrimeField, C, const M: usize, const A: usize, L>
+        StoreIntoCells<F, C, ExtractionSupport, L> for Base64Vec<F, M, A>
+    {
+        fn store(
+            self,
+            ctx: &mut OCtx<F, ExtractionSupport>,
+            chip: &C,
+            layouter: &mut impl LayoutAdaptor<F, ExtractionSupport, Adaptee = L>,
+            injected_ir: &mut IR<F>,
+        ) -> Result<(), Error> {
+            self.0.store(ctx, chip, layouter, injected_ir)
+        }
+    }
+}
+
 /// Equivalent to Base64Instructions for variable-length inputs.
 pub trait Base64VarInstructions<F: PrimeField, const M: usize, const A: usize>:
     Base64Instructions<F>